Change the location of the js and json folders
The js and json folders are now in the root directory, because these
folders and its content does not belong to the purpose of the assets
folder.

Also the resource parameter for the fetch method was modified in all the
instances were is used, because this method search for the resource
starting from the root directory.